+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 
 - Report invalid property names in `mappedBy` of `@OneToOne`, `@OneToMany` and `@ManyToMany`.
 - Complete valid property names in `mappedBy` of `@OneToOne`, `@OneToMany` and `@ManyToMany`.
+- Ctrl+Click on the `mappedBy` value in `@OneToOne`, `@OneToMany` and `@ManyToMany` now navigates to the referenced property.